Refactor LoggerHolder
