+++ examples/SettingUpLogging.php
@@ -6,7 +6,7 @@ require __DIR__ . '/../vendor/autoload.php';
 use Monolog\Handler\StreamHandler;
 use Monolog\Logger;
 use OpenTelemetry\Contrib\OtlpGrpc\Exporter as OtlpGrpcExporter;
-use OpenTelemetry\SDK\GlobalLoggerHolder;
+use OpenTelemetry\SDK\Common\Log\LoggerHolder;
 use OpenTelemetry\SDK\Trace\SpanProcessor\SimpleSpanProcessor;
 use OpenTelemetry\SDK\Trace\TracerProvider;
 use Psr\Log\LogLevel;
@@ -15,8 +15,9 @@ echo 'Starting SettingUpLogging example' . PHP_EOL;
 
 //create a Logger, and register it with the global logger holder. The library will use this logger
 //for all of its internal logging (errors, warnings, etc)
-$logger = new Logger('otel-php', [new StreamHandler(STDOUT, LogLevel::DEBUG)]);
-GlobalLoggerHolder::set($logger);
+LoggerHolder::set(
+    new Logger('otel-php', [new StreamHandler(STDOUT, LogLevel::DEBUG)])
+);
 
 $tracerProvider =  new TracerProvider(
     new SimpleSpanProcessor(
